requirements.txt
telebot~=0.0.5
requests~=2.32.4
bs4~=0.0.2
lxml
beautifulsoup4~=4.13.4
python-dotenv~=1.0.1
